\name{org.Slycopersicum.egCHR}
\alias{org.Slycopersicum.egCHR}
\title{Map between Entrez Gene IDs to Chromosomes}
\description{ 
  org.Slycopersicum.egCHR is an R object that provides mappings between
  entrez gene identifiers and the chromosome that contains the gene of interest.
}
\details{
  foo
}
\examples{
  dir()
}
\keyword{datasets}
