\name{org.Slycopersicum.egENZYME}
\alias{org.Slycopersicum.egENZYME}
\alias{org.Slycopersicum.egENZYME2ORF}
\title{Map between Entrez Gene IDs and Enzyme Commission (EC) Numbers}
\description{ 
  org.Slycopersicum.egENZYME is an R object that provides mappings between
  entrez gene identifiers and EC numbers.
}
\details{
  foo
}
\examples{
  dir()
}
\keyword{datasets}
