\name{org.Slycopersicum.egITAG}
\alias{org.Slycopersicum.egITAG}
\alias{org.Slycopersicum.egITAG2ORF}
\title{Map between Entrez Gene IDs and ITAG}
\description{
  org.Slycopersicum.egITAG is an R object that maps entrez gene
  identifiers to the corresponding ITAG. 
}
\details{
 
}

\examples{
}
\keyword{datasets}

