\name{org.Slycopersicum.egLYCOCYCENZYME}
\alias{org.Slycopersicum.egLYCOCYCENZYME}
\title{Map between Entrez Gene identifiers and Enzyme Names from LYCOCYC}
\description{ 
  org.Slycopersicum.egLYCOCYCENZYME is an R object that provides mappings between
  Entrez Gene identifiers and Enzyme Names from LYCOCYC. 
}
\details{
  foo
}
\examples{
  dir()
}
\keyword{datasets}

