\name{org.Slycopersicum.egORGANISM}
\alias{org.Slycopersicum.egORGANISM}
\title{The Organism for org.Slycopersicum.eg}
\description{
  org.Slycopersicum.egORGANISM is an R object that contains a single item: a
  character string that names the organism for which org.Slycopersicum.eg was built.
}
\details{
  Although the package name is suggestive of the organism for which it
  was built, org.Slycopersicum.egORGANISM provides a simple way to programmatically
  extract the organism name.
}
\examples{
  org.Slycopersicum.egORGANISM
}
\keyword{datasets}

