\name{org.Slycopersicum.egSGN}
\alias{org.Slycopersicum.egSGN}
\alias{org.Slycopersicum.egSGN2ORF}
\title{Map between Entrez Gene IDs and SGN}
\description{
  org.Slycopersicum.egSGN is an R object that maps entrez gene
  identifiers to the corresponding SGN. 
}
\details{
 
}

\examples{
}
\keyword{datasets}

