\name{org.Slycopersicum.egUNIPROT}
\alias{org.Slycopersicum.egUNIPROT}
\title{Map between Entrez Gene IDs and Uniprot}
\description{
  org.Slycopersicum.egUNIPROT is an R object that maps entrez gene
  identifiers to the corresponding Uniprot. 
}
\details{
  foo
}

\examples{
  dir()
}
\keyword{datasets}

