\name{org.Slycopersicum.egLYCOCYC}
\alias{org.Slycopersicum.egLYCOCYC}
\alias{org.Slycopersicum.egLYCOCYC2ORF}
\title{Mappings between Entrez Gene identifiers and LYCOCYC}
\description{
}
\details{
}
\references{
  \url{http://pathway.gramene.org/gramene/lycocyc.shtml}
}
\examples{
}
\keyword{datasets}

